custom hooks for task management
